+++ sdk/@launchdarkly/observability/CHANGELOG.md
@@ -1,5 +1,32 @@
 # @launchdarkly/observability
 
+## [1.0.3](https://github.com/launchdarkly/observability-sdk/compare/observability-1.0.2...observability-1.0.3) (2026-02-26)
+
+
+### Dependencies
+
+* The following workspace dependencies were updated
+  * dependencies
+    * highlight.run bumped to 9.27.1
+
+## [1.0.2](https://github.com/launchdarkly/observability-sdk/compare/observability-1.0.1...observability-1.0.2) (2026-02-23)
+
+
+### Dependencies
+
+* The following workspace dependencies were updated
+  * dependencies
+    * highlight.run bumped to 9.27.0
+
+## [1.0.1](https://github.com/launchdarkly/observability-sdk/compare/observability-1.0.0...observability-1.0.1) (2026-02-13)
+
+
+### Dependencies
+
+* The following workspace dependencies were updated
+  * dependencies
+    * highlight.run bumped to 9.26.1
+
 ## [1.0.0](https://github.com/launchdarkly/observability-sdk/compare/observability-0.5.2...observability-1.0.0) (2026-01-28)
 
 This release marks the 1.0.0 stable version of the LaunchDarkly Observability Browser SDK.